+++ src/antidiag.cpp
@@ -48,16 +48,27 @@ NumericMatrix antidiag_matrix(NumericVector x,
 
 //' @name antidiag
 //' @title Matrix antidiagonals
+//' @usage
+//' antidiag(x = 1, nrow, ncol)
+//' antidiag(x) <- value
 //' @description
 //' Extract or replace the antidiagonal of a matrix,
 //' or construct a antidiagonal matrix.
 //' @param x matrix, vector or 1D array, or missing.
 //' @param ncol number of columns (optional; when x is not a matrix)
 //' @param nrow number of rows (optional; when x is not a matrix)
+//' @param value either a single value or a vector of length equal to that of the current antidiagonal.
+//' Should be of a mode which can be coerced to that of x.
 //' @examples
+//' # Extracting antidiag
 //' antidiag(diag(3))
+//' # Creating antidiagonal matrix
 //' antidiag(7, 3, 3)
 //' antidiag(1:5, 3, 3)
+//' # Assigning antidiagonal
+//' mat <- matrix(0, 3, 3)
+//' antidiag(mat) <- c(3, 4, 5)
+//' mat
 //' @export
 // [[Rcpp::export]]
 RObject antidiag(RObject x = NumericVector::create(1),
@@ -71,6 +82,23 @@ RObject antidiag(RObject x = NumericVector::create(1),
   } else {
     return R_NilValue;
   }
+}
+
+//' @rdname antidag
+//' @export
+// [[Rcpp::export(name = `antidiag<-`)]]
+RObject antidiag_assign(NumericMatrix x, NumericVector value){
+  // https://stackoverflow.com/questions/58877742/is-there-any-way-in-which-to-make-an-infix-function-using-sourcecpp
+  // TODO: Funtions behaviour for long vectors
+  int n_row = x.nrow();
+  int n_col = x.ncol();
+
+  for(int i=0; i < n_row; i++){
+    int index = i % value.length();
+    x(i, n_row - 1 - i) = value(index);
+  }
+
+  return x;
 }