+++ NEWS.md
@@ -1,5 +1,8 @@
-# matricks 0.3.0
+# matricks 0.4.0
+* `antidiag<-` function added
+* `neighbour_idx` function added
 
+# matricks 0.3.0  
 * `NEWS.md` file added to track changes to the package.
 * `antidiag()` function bugfix
 * `row_bind` function added